break into smaller compoonents, grid layout for trail list
