+++ package.json
@@ -44,6 +44,7 @@
     "prop-types": "^15.6.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.12.0",
     "react-redux": "^8.1.3",
     "redux": "^4.0.5",
     "redux-mock-store": "^1.5.4"
